+++ x11/x11_client_container.h
@@ -17,14 +17,13 @@ public:
     X11ClientContainer();
     virtual ~X11ClientContainer();
 
-//     virtual void setFocus();
     virtual void setMapped(bool mapped);
     virtual void setRect(const Rect &rect);
     virtual void handleMaximizedChanged();
     virtual void handleActiveChanged() ;
     virtual int numElements() { return _children.size(); }
     virtual bool isEmpty() { return _children.empty(); }
-    virtual Client *child(int index) { return _children[index]; }
+    virtual Client *child(size_t index) { return _children[index]; }
     virtual int indexOfChild(const Client *child);
     virtual int activeChildIndex() { return _active_child_index; }
     virtual void setActiveChild(int index);
